Rework macOS icons
to support automatic switching with light and dark mode. This is known
as a template icon. It generally makes the icons fit better with the
macOS standard keyboard layout icons. Though, we are restricted to
a square icon. There seems to be no way of getting a rectangular item as
used in the most recent versions of macOS.

Additionally, stroke widths has been increased to make the font a bit
more legible in the small icons.

The new icons only contain sizes up to 32x32@2x. I don't think the
keyboard layout icons are displayed with a larger size anywhere, so this
should be sufficient (not even sure that the 32x32 icon is used).