Update layer mod3 special characters
They were messed up when I simply took the Bone layout and opened it in Ukelele and swapped the full key definitions so that the letters are on the correct position. I thought that all mod3 layer definitions are in the Karabiner Elements config file, but no. This keylayout file here containts those!

I now used Ukelele and edited in its UI the definitions of the mod3 layer.

The code is a bit inconsistent now. In the next commit I will try to fix those few cases by editing the file by hand.