Fix NeoQwertz layer mod4 the key "Ü"
This was broken in the published version. The key with id="33" or in Karabiner langluage key_code="open_bracket", where in German ISO layout there is an "Ü"
This was broken for the Mod4 layer with the navigation and virtual numpad.
It produced an "Ü" instead of a rightfully "−".
I found out that this problem does not come from Karabiner complex modification rules, but from this keylayout file.
The Mod4 layer is crazy. The Karabiner rules define the layer switching key, but the layer definitions are split between the karabiner rules and this file. 🤪

Now this button is fixed. Probably no one noticed it because it's seldom used.